Split out sound system into its own file
