+++ src/platforms/sdl3/platform.c
@@ -1,7 +1,6 @@
 #include <stdbool.h>
 
 #include <SDL3/SDL.h>
-#include <SDL3_mixer/SDL_mixer.h>
 #include <SDL3/SDL_error.h>
 
 #include "../extensions/sdl3-extensions.h"
@@ -16,7 +15,6 @@
 #include "../../log.h"
 #include "../../math.h"
 #include "../../platform.h"
-#include "../../sounds.h"
 
 #include "../../modules/platforms/sdl3.h"
 
@@ -30,25 +28,16 @@ static uint32_t* pixels = NULL;
 static int ticks_last_frame;
 static SDL_FRect display_rect;
 
-#define MIX_CHANNELS 8
-static MIX_Track* tracks[MIX_CHANNELS];
-static bool audio_disabled = false;
-static MIX_Mixer* mixer = NULL;
-
-SDL_AudioSpec spec = {
-    .format = SDL_AUDIO_U8,
-    .channels = 1,
-    .freq = 11025
-};
-
 static void sdl_init(void);
 static void sdl_destroy(void);
 static void sdl_handle_events(void);
 static void sdl_fix_frame_rate(void);
 static void sdl_pixels_resize(int width, int height);
 
-static void mixer_init(void);
-static void mixer_destroy(void);
+void sound_init(void);
+void sound_reload(void);
+void sound_destroy(void);
+void sound_get_version(char* s);
 
 static void log_platform_info(void);
 
@@ -67,19 +56,18 @@ int platform_main(int argc, char* argv[]) {
 }
 
 void platform_init(void) {
-    mixer_init();
+    sound_init();
     sdl_init();
     log_platform_info();
 }
 
 void platform_destroy(void) {
-    mixer_destroy();
+    sound_destroy();
     sdl_destroy();
 }
 
 void platform_reload(void) {
-    MIX_StopAllTracks(mixer, 0);
-    //MIX_Volume(-1, MIX_MAX_VOLUME);
+    sound_reload();
 }
 
 void platform_update(void) {
@@ -166,70 +154,6 @@ bool platform_handle_event(event_t* event) {
     }
 
     return false;
-}
-
-void platform_sound_play(sound_t* sound, int channel, bool looping) {
-    if (audio_disabled) return;
-
-    if (channel >= MIX_CHANNELS) {
-        log_error("Error playing sound: channel %i does not exist", channel);
-        return;
-    }
-
-    // Search for a free channel if channel not specified
-    if (channel == -1) {
-        for (int i = 0; i < MIX_CHANNELS; i++) {
-            MIX_Track* track = tracks[i];
-            if (!MIX_TrackPlaying(track)) {
-                channel = i;
-                break;
-            }
-        }
-    }
-
-    if (channel == -1) {
-        log_error("Error playing sound: no free channels available");
-        return;
-    }
-
-    size_t size = sound->frame_count * sound->channel_count * sizeof(sample_t);
-
-    MIX_Audio* audio = MIX_LoadRawAudioNoCopy(
-        mixer,
-        sound->pcm,
-        size,
-        &spec,
-        false
-    );
-
-    int loops = looping ? -1 : 0;
-    MIX_Track* track = tracks[channel];
-
-    MIX_SetTrackAudio(track, audio);
-    MIX_PlayTrack(track, 0);
-    MIX_SetTrackLoops(track, loops);
-}
-
-void platform_sound_stop(int channel) {
-    if (channel < -1 || channel >= MIX_CHANNELS) {
-        log_error("Error stopping channel: channel %i does not exist", channel);
-        return;
-    }
-
-    MIX_Track* track = tracks[channel];
-    MIX_StopTrack(track, 0);
-}
-
-void platform_sound_volume(int channel, float volume) {
-    if (channel < -1 || channel >= MIX_CHANNELS) {
-        log_error("Error stopping channel: channel %i does not exist", channel);
-        return;
-    }
-
-    volume = clamp(volume, 0.0f, 1.0f);
-
-    MIX_Track* track = tracks[channel];
-    MIX_SetTrackGain(track, volume);
 }
 
 void platform_mouse_grabbed_set(bool grabbed) {
@@ -450,79 +374,24 @@ static void sdl_pixels_resize(int width, int height) {
     }
 }
 
-static void mixer_init(void) {
-    // Validate SDL version
-    const int sdl_version = SDL_GetVersion();
-    const int version_major = SDL_VERSIONNUM_MAJOR(sdl_version);
-    const int version_minor = SDL_VERSIONNUM_MINOR(sdl_version);
-    if (version_major != 3 && version_minor < 4) {
-        log_fatal("SDL_mixer 3 requires SDL 3.4.0 or later");
-    }
-
-    if (!MIX_Init()) {
-        log_error("Error intializing SDL Mixer");
-        log_error(SDL_GetError());
-        log_info("Sound playback will be disabled");
-        audio_disabled = true;
-        return;
-    }
-
-    // Create mixer device
-    mixer = MIX_CreateMixerDevice(SDL_AUDIO_DEVICE_DEFAULT_PLAYBACK, &spec);
-    if (!mixer) {
-        log_error("Error creating mixer device");
-        log_error(SDL_GetError());
-        log_info("Sound playback will be disabled");
-        audio_disabled = true;
-        return;
-    }
-
-    // Create mixer tracks
-    for (int i = 0; i < MIX_CHANNELS; i++) {
-        tracks[i] = MIX_CreateTrack(mixer);
-
-        if (!tracks[i]) {
-            log_error("Error creating track %i", i);
-            log_error(SDL_GetError());
-            log_info("Sound playback will be disabled");
-            audio_disabled = true;
-            return;
-        }
-    }
-}
-
-static void mixer_destroy(void) {
-    MIX_StopAllTracks(mixer, 0);
-
-    for (int i = 0; i < MIX_CHANNELS; i++) {
-        MIX_DestroyTrack(tracks[i]);
-        tracks[i] = NULL;
-    }
-
-    MIX_DestroyMixer(mixer);
-
-    MIX_Quit();
-}
-
 static void log_platform_info(void) {
     // Get SDL version info
     const int sdl_version = SDL_GetVersion();
 
-    // Get SDL Mixer version info
-    const int mix_version = MIX_Version();
+    // Get sound system version info
+    char sound_version[32];
+    sound_get_version(sound_version);
 
     // Build info string
     char buffer[128];
     snprintf(
         buffer,
         sizeof(buffer),
-        "platform init (SDL %i.%i.%i, SDL Mixer %i.%i.%i)",
+        "platform init (SDL %i.%i.%i, %s)",
         SDL_VERSIONNUM_MAJOR(sdl_version),
         SDL_VERSIONNUM_MINOR(sdl_version),
         SDL_VERSIONNUM_MICRO(sdl_version),
-        SDL_VERSIONNUM_MAJOR(mix_version),
-        SDL_VERSIONNUM_MINOR(mix_version),
-        SDL_VERSIONNUM_MICRO(mix_version)
+        sound_version
     );
 
     log_info(buffer);
